chore: extract useCurrentAppState hook to reduce cyclic dependency
