+++ app/client/src/pages/Editor/IDE/Layout/hooks/useGridLayoutTemplate.ts
@@ -2,7 +2,8 @@ import React from "react";
 import type { AnimatedGridUnit } from "components/AnimatedGridLayout";
 import { useSelector } from "react-redux";
 import useWindowDimensions from "utils/hooks/useWindowDimensions";
-import { useCurrentAppState, useCurrentEditorState } from "../../hooks";
+import { useCurrentEditorState } from "../../hooks";
+import { useCurrentAppState } from "../../hooks/useCurrentAppState";
 import { getPropertyPaneWidth } from "selectors/propertyPaneSelectors";
 import { previewModeSelector } from "selectors/editorSelectors";
 import { getIDEViewMode } from "selectors/ideSelectors";
