+++ app/client/src/pages/Editor/WidgetsEditor/WidgetEditorContainer.tsx
@@ -2,7 +2,7 @@ import EditorContextProvider from "components/editorComponents/EditorContextProv
 import type { ReactNode } from "react";
 import React from "react";
 import classNames from "classnames";
-import { useCurrentAppState } from "pages/Editor/IDE/hooks";
+import { useCurrentAppState } from "pages/Editor/IDE/hooks/useCurrentAppState";
 import { useSelector } from "react-redux";
 import { getIsAppSettingsPaneWithNavigationTabOpen } from "selectors/appSettingsPaneSelectors";
 import { EditorState } from "ee/entities/IDE/constants";
