+++ app/client/src/pages/Editor/WidgetsEditor/components/NavigationAdjustedPageViewer.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode } from "react";
 import React from "react";
 import { EditorState } from "ee/entities/IDE/constants";
-import { useCurrentAppState } from "pages/Editor/IDE/hooks";
+import { useCurrentAppState } from "pages/Editor/IDE/hooks/useCurrentAppState";
 import { getIsAppSettingsPaneWithNavigationTabOpen } from "selectors/appSettingsPaneSelectors";
 import { useSelector } from "react-redux";
 import { combinedPreviewModeSelector } from "selectors/editorSelectors";
