+++ app/client/src/pages/Editor/WidgetsEditor/components/WidgetEditorNavigation.tsx
@@ -1,7 +1,7 @@
 import React, { forwardRef, useEffect, useRef, useState } from "react";
 import NavigationPreview from "./NavigationPreview";
 import { EditorState } from "ee/entities/IDE/constants";
-import { useCurrentAppState } from "pages/Editor/IDE/hooks";
+import { useCurrentAppState } from "pages/Editor/IDE/hooks/useCurrentAppState";
 import {
   getAppSettingsPaneContext,
   getIsAppSettingsPaneWithNavigationTabOpen,
